+++ src/components/hooks/location-hooks.js
@@ -1,12 +1,25 @@
 import React from 'react'
+import { useWebsiteStatus } from './website-status-hooks'
+import { CookieStorage } from 'common/storage'
+import { isEuCountry } from 'common/country-base'
 
-// Fires a callback after a useState update
-export const useStateWithCallback = (initialState, callback) => {
-    const [state, setState] = React.useState(initialState)
+const CLIENTS_COUNTRY_KEY = 'clients_country'
+const clients_country_cookie = new CookieStorage(CLIENTS_COUNTRY_KEY)
+
+export const useEuCountry = () => {
+    const [clients_country, setClientCountry] = React.useState(
+        clients_country_cookie.get(CLIENTS_COUNTRY_KEY),
+    )
+    const [is_eu_country, setEuCountry] = React.useState(isEuCountry(clients_country))
+    const [website_status] = useWebsiteStatus()
 
     React.useEffect(() => {
-        callback(state)
-    }, [state, callback])
+        if (!clients_country && website_status) {
+            setClientCountry(website_status.clients_country)
+            clients_country_cookie.set(CLIENTS_COUNTRY_KEY, website_status.clients_country)
+            setEuCountry(isEuCountry(website_status.clients_country))
+        }
+    }, [website_status])
 
-    return [state, setState]
+    return [is_eu_country]
 }
